Christian testing

+++ Half-step/include/motorMovementFunctions.h
@@ -10,4 +10,4 @@ void move_Center_From_X2();
 void make_step_Y(_Bool direction);
 void make_step_X(_Bool direction);
 void move_full_circle(int radius);
-void move_deltas(double dx, double dy);
+void move_deltas(int dx, int dy);
